# docs/openapi/swagger.json
{
    "components": {
        "schemas": {
...
            "IssueResponse": {
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/digest.Issue"
                    },
                    "error": {
                        "type": "boolean"
                    },
                    "message": {
                        "example": "Successfully retrieved issue",
                        "type": "string"
                    },
                    "status": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "ItemResponse": {
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/news.Item"
                    },
                    "error": {
                        "type": "boolean"
                    },
                    "message": {
                        "example": "Successfully retrieved item",
                        "type": "string"
                    },
                    "status": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "Response": {
                "properties": {
                    "data": {},
                    "error": {
                        "type": "boolean"
                    },
                    "message": {
                        "example": "User formatted message from the API",
                        "type": "string"
                    },
                    "status": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "digest.Issue": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "items": {
                        "items": {
                            "$ref": "#/components/schemas/news.Item"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "sent_at": {
                        "type": "string"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "status": {
                        "$ref": "#/components/schemas/digest.IssueStatus"
                    },
                    "subject": {
                        "type": "string"
                    },
                    "summary": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "digest.IssueStatus": {
                "enum": [
                    "draft",
                    "sent",
                    "error"
                ],
                "type": "string",
                "x-enum-varnames": [
                    "IssueStatusDraft",
                    "IssueStatusSent",
                    "IssueStatusError"
                ]
            },
...
            "news.Author": {
                "properties": {
                    "avatar_url": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "profile_url": {
                        "type": "string"
                    },
                    "username": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "news.Item": {
                "properties": {
                    "author": {
                        "$ref": "#/components/schemas/news.Author"
                    },
                    "comments": {
                        "type": "integer"
                    },
                    "id": {
                        "type": "integer"
                    },
                    "image_url": {
                        "type": "string"
                    },
                    "original_url": {
                        "description": "listing on the source platform (e.g. HN comments page), when different from URL",
                        "type": "string"
                    },
                    "published": {
                        "type": "string"
                    },
                    "score": {
                        "description": "per-source relevance/popularity, normalised across sources",
                        "type": "number"
                    },
                    "snippet": {
                        "type": "string"
                    },
                    "source": {
                        "$ref": "#/components/schemas/news.Source"
                    },
                    "tag": {
                        "$ref": "#/components/schemas/news.Tag"
                    },
                    "title": {
                        "type": "string"
                    },
                    "url": {
                        "description": "click target — the external content the source is linking to",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "news.Source": {
                "enum": [
                    "dev_to",
                    "go_blog",
                    "github",
                    "github_trending",
                    "reddit",
                    "hacker_news",
                    "golangbridge",
                    "lobsters",
                    "medium",
                    "youtube",
                    "go_podcast",
                    "fallthrough",
                    "ardanlabs_podcast",
                    "go_release",
                    "mastodon",
                    "awesome_go",
                    "jetbrains",
                    "golang_nuts",
                    "planet_golang",
                    "meetup",
                    "conferences",
                    "go_vuln",
                    "freecodecamp",
                    "hacker_news_jobs",
                    "remote_ok"
                ],
                "type": "string",
                "x-enum-varnames": [
                    "SourceDevTo",
                    "SourceGoBlog",
                    "SourceGitHub",
                    "SourceGitHubTrending",
                    "SourceReddit",
                    "SourceHN",
                    "SourceGolangBridge",
                    "SourceLobsters",
                    "SourceMedium",
                    "SourceYouTube",
                    "SourceGoPodcast",
                    "SourceFallthrough",
                    "SourceArdanLabs",
                    "SourceGoRelease",
                    "SourceMastodon",
                    "SourceAwesomeGo",
                    "SourceJetBrains",
                    "SourceGolangNuts",
                    "SourcePlanetGolang",
                    "SourceMeetup",
                    "SourceConferences",
                    "SourceGoVuln",
                    "SourceFreeCodeCamp",
                    "SourceHNJobs",
                    "SourceRemoteOK"
                ]
            },
            "news.Tag": {
                "description": "source-specific hint (\"proposal-accepted\", \"trending\", \"official\")",
                "enum": [
                    "article",
                    "tutorial",
                    "proposal",
                    "proposal_accepted",
                    "proposal_shipped",
                    "video",
                    "podcast",
                    "release",
                    "security",
                    "discussion",
                    "trending",
                    "event",
                    "conference",
                    "conference_reminder",
                    "conference_alert",
                    "jobs"
                ],
                "type": "string",
                "x-enum-comments": {
                    "TagConference": "major Go conference announcement",
                    "TagConferenceAlert": "~1 week before",
                    "TagConferenceReminder": "~3 months before"
                },
                "x-enum-varnames": [
                    "TagArticle",
                    "TagTutorial",
                    "TagProposal",
                    "TagProposalAccepted",
                    "TagProposalShipped",
                    "TagVideo",
                    "TagPodcast",
                    "TagRelease",
                    "TagSecurity",
                    "TagDiscussion",
                    "TagTrending",
                    "TagEvent",
                    "TagConference",
                    "TagConferenceReminder",
                    "TagConferenceAlert",
                    "TagJobs"
                ]
            }
        },
        "securitySchemes": {
            "BearerAuth": {
                "description": "Bearer token supplied in the Authorization header.",
                "in": "header",
                "name": "Authorization",
                "type": "apiKey"
            }
        }
    },
    "info": {
        "contact": {
            "name": "ainsley.dev",
            "url": "https://ainsley.dev"
        },
        "description": "The GoDaily news digest API — ranked, summarised Go news delivered daily.",
        "license": {
            "name": "BSD-style"
        },
        "title": "GoDaily API",
        "version": "1.0"
    },
...
    "paths": {
...
        "/issues/{slug}": {
            "get": {
                "description": "Returns a single digest issue identified by its date slug.",
                "parameters": [
                    {
                        "description": "Issue date slug",
                        "in": "path",
                        "name": "slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IssueResponse"
                                }
                            }
                        },
                        "description": "Successfully retrieved issue"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Response"
                                }
                            }
                        },
                        "description": "Slug is required"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Response"
                                }
                            }
                        },
                        "description": "Issue not found"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Response"
                                }
                            }
                        },
                        "description": "Failed to fetch issue"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Fetch an issue by slug.",
                "tags": [
                    "issues"
                ]
            }
        },
        "/items/{id}": {
            "get": {
                "description": "Returns a single news item identified by its numeric ID.",
                "parameters": [
                    {
                        "description": "Item ID",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ItemResponse"
                                }
                            }
                        },
                        "description": "Successfully retrieved item"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Response"
                                }
                            }
                        },
                        "description": "ID is required or not a positive integer"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Response"
                                }
                            }
                        },
                        "description": "Item not found"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Response"
                                }
                            }
                        },
                        "description": "Failed to fetch item"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Fetch a news item by ID.",
                "tags": [
                    "items"
                ]
            }
        },
...
    },
    "openapi": "3.1.0",
    "servers": [
        {
            "url": "/api"
        }
    ]
}